Rename the python package to "zabbix-smartmontools"
