+++ Win/zabbix_agentd.conf.d/userparameter_smartctl.conf
@@ -1,1 +1,1 @@
-UserParameter=smartctl.discovery[*], python "C:\I\dont\know\the\correct\path\smartctl-lld.exe" "$1" "$2"
+UserParameter=smartctl.discovery[*], python "C:\I\dont\know\the\correct\path\zabbix-smartctl.exe" "$1" "$2"
