+++ zabbix_smartmontools/sender_wrapper.py
